Fix bug missing MapView after importing Map
