+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/schema/Property.java
@@ -171,6 +171,11 @@ public class Property implements Ordered, Klass
     private boolean embeddedObject;
 
     /**
+     * Does this class implement {@link EmbeddedObject} ?
+     */
+    private boolean analyticalObject;
+
+    /**
      * Can this property be read.
      */
     private boolean readable;
@@ -526,6 +531,18 @@ public class Property implements Ordered, Klass
     public void setEmbeddedObject( boolean embeddedObject )
     {
         this.embeddedObject = embeddedObject;
+    }
+
+    @JsonProperty
+    @JacksonXmlProperty( namespace = DxfNamespaces.DXF_2_0 )
+    public boolean isAnalyticalObject()
+    {
+        return analyticalObject;
+    }
+
+    public void setAnalyticalObject(boolean analyticalObject)
+    {
+        this.analyticalObject = analyticalObject;
     }
 
     @JsonProperty
