+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/deletedobject/hibernate/DeletedObjectPostDeleteEventListener.java
@@ -31,6 +31,7 @@ package org.hisp.dhis.deletedobject.hibernate;
 import org.hibernate.event.spi.PostDeleteEvent;
 import org.hibernate.event.spi.PostDeleteEventListener;
 import org.hibernate.persister.entity.EntityPersister;
+import org.hisp.dhis.common.EmbeddedObject;
 import org.hisp.dhis.common.IdentifiableObject;
 import org.hisp.dhis.common.MetadataObject;
 import org.hisp.dhis.common.UserContext;
@@ -44,7 +45,7 @@ public class DeletedObjectPostDeleteEventListener implements PostDeleteEventList
     @Override
     public void onPostDelete( PostDeleteEvent event )
     {
-        if ( MetadataObject.class.isInstance( event.getEntity() ) )
+        if ( MetadataObject.class.isInstance( event.getEntity() ) && !EmbeddedObject.class.isInstance( event.getEntity() ) )
         {
             IdentifiableObject identifiableObject = (IdentifiableObject) event.getEntity();
             DeletedObject deletedObject = new DeletedObject( identifiableObject );
