+++ dhis-2/dhis-services/dhis-service-reporting/src/main/java/org/hisp/dhis/mapping/MapDeletionHandler.java
@@ -59,17 +59,4 @@ public class MapDeletionHandler
     {
         return Map.class.getSimpleName();
     }
-
-    @Override
-    public void deleteMap( Map map )
-    {
-        Iterator<MapView> views = map.getMapViews().iterator();
-
-        while ( views.hasNext() )
-        {
-            MapView view = views.next();
-            views.remove();
-            mappingService.deleteMapView( view );
-        }
-    }
 }
