+++ dhis-2/dhis-services/dhis-service-reporting/src/test/resources/create_map.json
@@ -1,0 +1,74 @@
+{
+  "dataElements": [
+    {
+      "code": "DECA",
+      "created": "2016-03-04T03:37:12.714+0000",
+      "lastUpdated": "2016-03-04T03:37:12.733+0000",
+      "name": "ANC 1st visit",
+      "id": "fbfJHSPpUQD",
+      "shortName": "ANC 1st visit",
+      "aggregationType": "SUM",
+      "domainType": "AGGREGATE",
+      "publicAccess": "--------",
+      "description": "DEDA",
+      "zeroIsSignificant": false,
+      "valueType": "INTEGER",
+      "userGroupAccesses": [],
+      "attributeValues": [],
+      "aggregationLevels": []
+    }
+  ],
+  "maps": [
+    {
+      "basemap": "osmLight",
+      "mapViews": [
+        {
+          "id" : "zyFOjTfzLws",
+          "columns": [
+            {
+              "dimension": "dx",
+              "items": [
+                {
+                  "id": "fbfJHSPpUQD",
+                  "name": "ANC 1st visit",
+                  "dimensionItemType": "DATA_ELEMENT"
+                }
+              ],
+              "objectName": "de"
+            }
+          ],
+          "endDate": "2019-09-01",
+          "filters": [
+            {
+              "dimension": "pe",
+              "items": [
+                {
+                  "id": "LAST_12_MONTHS",
+                  "name": "Last 12 months"
+                }
+              ]
+            }
+          ],
+          "layer": "thematic",
+          "name": "ANC 1st visit",
+          "opacity": 0.9,
+          "rows": [
+            {
+              "dimension": "ou",
+              "items": [
+                {
+                  "id": "LEVEL-2"
+                }
+              ]
+            }
+          ],
+          "startDate": "2018-09-01",
+          "valueType": "de"
+        }
+      ],
+      "name": "test1",
+      "description": "test",
+      "id": "LTNgXfzTFTv"
+    }
+  ]
+}
