+++ dhis-2/dhis-services/dhis-service-schema/src/main/java/org/hisp/dhis/schema/Jackson2PropertyIntrospectorService.java
@@ -37,6 +37,7 @@ import com.google.common.collect.Multimap;
 import com.google.common.primitives.Primitives;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.hisp.dhis.common.AnalyticalObject;
 import org.hisp.dhis.common.IdentifiableObject;
 import org.hisp.dhis.common.EmbeddedObject;
 import org.hisp.dhis.common.NameableObject;
@@ -197,6 +198,7 @@ public class Jackson2PropertyIntrospectorService
                     property.setIdentifiableObject( IdentifiableObject.class.isAssignableFrom( klass ) );
                     property.setNameableObject( NameableObject.class.isAssignableFrom( klass ) );
                     property.setEmbeddedObject( EmbeddedObject.class.isAssignableFrom( klass ) );
+                    property.setAnalyticalObject( AnalyticalObject.class.isAssignableFrom( klass ) );
                 }
             }
             else
